+++ epa/src/main/kotlin/moritz/lindner/masterarbeit/epa/api/EpaService.kt
@@ -1,12 +1,9 @@
 package moritz.lindner.masterarbeit.epa.api
 
-import io.github.oshai.kotlinlogging.KotlinLogging
 import moritz.lindner.masterarbeit.epa.ExtendedPrefixAutomaton
+import moritz.lindner.masterarbeit.epa.domain.Event
+import moritz.lindner.masterarbeit.epa.features.animation.EventsByCasesCollector
 import moritz.lindner.masterarbeit.epa.features.filter.EpaFilter
-import moritz.lindner.masterarbeit.epa.features.layout.TreeLayout
-import moritz.lindner.masterarbeit.epa.features.layout.factory.LayoutConfig
-import moritz.lindner.masterarbeit.epa.features.layout.factory.LayoutFactory
-import moritz.lindner.masterarbeit.epa.features.layout.tree.EpaToTree
 import moritz.lindner.masterarbeit.epa.features.statistics.Statistics
 import moritz.lindner.masterarbeit.epa.features.statistics.StatisticsVisitor
 
@@ -21,26 +18,10 @@ class EpaService<T : Comparable<T>> {
     fun applyFilters(epa: ExtendedPrefixAutomaton<T>, filters: List<EpaFilter<T>>): ExtendedPrefixAutomaton<T> {
         return filters.fold(epa) { acc, filter -> filter.apply(acc) }
     }
-}
 
-class LayoutService<T : Comparable<T>> {
-
-    private val logger = KotlinLogging.logger {}
-
-    // TODO: maybe add caching
-    fun buildLayout(
-        epa: ExtendedPrefixAutomaton<Long>,
-        layoutConfig: LayoutConfig
-    ): TreeLayout {
-        logger.info { "building tree" }
-        val treeVisitor = EpaToTree<Long>()
-        epa.copy().acceptDepthFirst(treeVisitor)
-
-        logger.info { "building tree layout" }
-        val layout = LayoutFactory.create(layoutConfig)
-
-        layout.build(treeVisitor.root)
-
-        return layout
+    fun getEventsByCase(epa: ExtendedPrefixAutomaton<T>): Map<String, List<Event<T>>> {
+        val visitor = EventsByCasesCollector<T>()
+        epa.acceptDepthFirst(visitor)
+        return visitor.build()
     }
 }
