+++ ui/src/main/kotlin/moritz/lindner/masterarbeit/ui/components/epaview/components/animation/SingleCaseAnimationUI.kt
@@ -26,7 +26,8 @@ import androidx.compose.ui.window.DialogWindow
 import kotlinx.coroutines.ExecutorCoroutineDispatcher
 import kotlinx.coroutines.withContext
 import moritz.lindner.masterarbeit.epa.ExtendedPrefixAutomaton
-import moritz.lindner.masterarbeit.epa.features.animation.EventsByCasesCollector
+import moritz.lindner.masterarbeit.epa.api.EpaService
+import moritz.lindner.masterarbeit.epa.domain.Event
 import moritz.lindner.masterarbeit.ui.components.epaview.state.AnimationState
 import moritz.lindner.masterarbeit.ui.components.epaview.viewmodel.EpaViewModel
 import org.jetbrains.jewel.foundation.theme.JewelTheme
@@ -40,17 +41,18 @@ fun SingleCaseAnimationUI(
     viewModel: EpaViewModel,
     onClose: () -> Unit,
 ) {
+    val epaService = EpaService<Long>()
+
     var isLoading by remember { mutableStateOf(true) }
     var showDialog by remember { mutableStateOf(false) }
     var selectedCase by remember { mutableStateOf<String?>(null) }
-    var eventsByCasesCollector by remember { mutableStateOf(EventsByCasesCollector<Long>()) }
+    var eventsByCasesCollector by remember { mutableStateOf(emptyMap<String, List<Event<Long>>>()) }
 
     LaunchedEffect(epa) {
         isLoading = true
         withContext(backgroundDispatcher) {
             selectedCase = null
-            eventsByCasesCollector = EventsByCasesCollector()
-            epa.copy().acceptDepthFirst(eventsByCasesCollector)
+            eventsByCasesCollector = epaService.getEventsByCase(epa)
         }
         isLoading = false
     }
@@ -108,7 +110,7 @@ fun SingleCaseAnimationUI(
                     ) {
                         Row {
                             LazyColumn {
-                                items(eventsByCasesCollector.cases.toList()) { case ->
+                                items(eventsByCasesCollector.keys.toList()) { case ->
                                     val isSelected = case == selectedCase
                                     Text(
                                         text = case,
